@@ -2,30 +2,27 @@ import 'package:bitalino/bitalino.dart';
 import 'package:flutter/cupertino.dart';
 
 class ExamSettings {
-  Frequency sampleFrequency = Frequency.HZ1000;
-  int secondsToShow = 2;
-  int duration = 15;
+  Frequency sampleFrequency;
+  int secondsToShow;
+  int duration;
 
-  ExamSettings({this.sampleFrequency, this.secondsToShow, this.duration});
+  ExamSettings({this.sampleFrequency = Frequency.HZ1000, this.secondsToShow = 2, this.duration = 15});
 }
 
 class SettingsViewModel extends ChangeNotifier {
-  final startDuration = 15;
-  final startSecondsToShow = 2;
-  ExamSettings examSettings;
+  final int minDuration = 15;
+  final int minSecondsToShow = 2;
+  late final ExamSettings examSettings;
 
   SettingsViewModel() {
-    examSettings = ExamSettings(
-        sampleFrequency: Frequency.HZ1000,
-        secondsToShow: this.startSecondsToShow,
-        duration: this.startDuration);
+    examSettings = ExamSettings();
   }
 
   List<Frequency> getPossibleFrequencies() {
     return [Frequency.HZ1, Frequency.HZ10, Frequency.HZ100, Frequency.HZ1000];
   }
 
-  void setSampleFrequency(String frequency) {
+  void setSampleFrequency(String? frequency) {
     List<Frequency> frequencies = getPossibleFrequencies();
     Frequency newFreq =
         frequencies.firstWhere((element) => element.toString() == frequency);
@@ -33,24 +30,24 @@ class SettingsViewModel extends ChangeNotifier {
     notifyListeners();
   }
 
-  void setDuration(String newDur) {
-    this.examSettings.duration = int.parse(newDur);
+  void setDuration(String? newDur) {
+    this.examSettings.duration = int.parse(newDur==null? minDuration.toString(): newDur);
     if (this.examSettings.duration < this.examSettings.secondsToShow)
       this.examSettings.secondsToShow = getPossibleTimesToShow().last;
     notifyListeners();
   }
 
-  void setSecondsToShow(String newDur) {
-    this.examSettings.secondsToShow = int.parse(newDur);
+  void setSecondsToShow(String? newSecondsToShow) {
+    this.examSettings.secondsToShow = int.parse(newSecondsToShow==null? minSecondsToShow.toString(): newSecondsToShow);
     notifyListeners();
   }
 
   List<int> getPossibleDurations() {
-    return List.generate(6, (index) => startDuration * (index + 1));
+    return List.generate(6, (index) => minDuration * (index + 1));
   }
 
   List<int> getPossibleTimesToShow() {
-    return List.generate(10, (index) => startSecondsToShow * (index + 1))
+    return List.generate(10, (index) => minSecondsToShow * (index + 1))
         .where((element) => element <= this.examSettings.duration)
         .toList();
   }
