@@ -1,11 +1,20 @@
 import 'dart:math';
 
+import 'package:scidart/src/numdart/arrays_base/array.dart';
+
+import '../filters.dart';
 import '../isensor.dart';
 import '../sensors_codes.dart';
 
 class EmgSensorData extends ISensor {
-  EmgSensorData() : super(TechnicalInfo(vcc: 3.3, measurementUnit: "mV"),
-      SensorDataInfo(maxValue: 1.7, minValue: -1.7, sensorGain: 1009));
+  EmgSensorData()
+      : super(
+            TechnicalInfo(vcc: 3.3, measurementUnit: "mV"),
+            SensorDataInfo(maxValue: 1.7, minValue: -1.7, sensorGain: 1009),
+            BandPassFilter(
+                sampleFrequency: 1000, //TODO change this to be dynamic
+                lowCutOff: 25,
+                highCutOff: 480));
 
   int differenceToUnit = 1000;
 
@@ -16,12 +25,10 @@ class EmgSensorData extends ISensor {
 
   @override
   double applyTransferFunction(double valueSampled) {
-    return ((((valueSampled / pow(2, this.sensorDataInfo.channelBits)) - (1 / 2)) *
-        this.technicalInfo.vcc) /
-        this.sensorDataInfo.sensorGain) *
+    return ((((valueSampled / pow(2, this.sensorDataInfo.channelBits)) -
+                    (1 / 2)) *
+                this.technicalInfo.vcc) /
+            this.sensorDataInfo.sensorGain) *
         differenceToUnit;
   }
-
-  @override
-  double applyFilter(double normalizedValue) => normalizedValue;
 }
