@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class BackTitleAppbar extends StatelessWidget {
+class BackTitleAppbar extends StatelessWidget implements PreferredSizeWidget {
   final String title;
 
   const BackTitleAppbar(this.title);
@@ -13,4 +13,7 @@ class BackTitleAppbar extends StatelessWidget {
       centerTitle: true,
     );
   }
+
+  @override
+  Size get preferredSize =>  Size.fromHeight(60.0);
 }
