@@ -1,5 +1,4 @@
 import 'package:flutter/cupertino.dart';
-import 'package:flutter_blue/flutter_blue.dart';
 import 'package:telemon_app/src/data/services/bluetooth_service/bluetooth_controller.dart';
 import 'package:telemon_app/src/data/services/bluetooth_service/bluetooth_dev.dart';
 import 'package:telemon_app/src/data/services/device_service/bitalino/bitalino_controller.dart';
@@ -12,12 +11,12 @@ class DeviceViewModel extends ChangeNotifier {
   BitalinoController _deviceController = BitalinoController();
   BluetoothController _bluetoothController = BluetoothController();
 
-  Future<void> connectDevice({String macAddress, String uuid}) async {
+  Future<void> connectDevice(BluetoothDev bluetoothDev) async {
     if (currDeviceMac.isEmpty) {
       //TODO uncomment for bitalino
-      await _deviceController.connectToDevice(uuid);
-    }
-    currDeviceMac = macAddress;
+      await _deviceController.connectToDevice(bluetoothDev);
+    }//TODO watch this
+    currDeviceMac = bluetoothDev.macId;
     notifyListeners();
   }
 
